Make Pageable helpers internal
